Update confirmation and on-hold pages to new styles
