--- mtp_send_money/assets-src/__old__/javascripts/modules/reference.js
@@ -1,19 +1,0 @@
-// Reference section
-'use strict';
-
-exports.Reference = {
-  init: function () {
-    $('.mtp-print-reference').click(function (e) {
-      e.preventDefault();
-      try {
-        window.print();
-      } catch (e) {}  // eslint-disable-line
-    });
-
-    $('.mtp-email-reference').click(function (e) {
-      e.preventDefault();
-      $('#mtp-email-reference-form').show();
-      $('#id_email').focus();
-    });
-  }
-};
