+++ mtp_send_money/assets-src/__old__/stylesheets/app.scss
@@ -15,45 +15,3 @@
 @import 'views/send-money';
 @import 'views/help-pages';
 @import 'views/filtered-list';
-
-// Govuk elements overrides
-
-.govuk-box-highlight {
-  padding: 2em;
-
-  p, strong {
-    @include core-36;
-  }
-
-  p {
-    margin-bottom: 2px;
-  }
-
-  strong {
-    font-weight: 700;
-  }
-
-  h1 {
-    margin-top: 0;
-    margin-bottom: 0.6em;
-  }
-
-  h2 {
-    margin-top: 0;
-    margin-bottom: 0;
-  }
-
-  &.mtp-box-highlight-info {
-    background: $govuk-blue;
-  }
-
-  &.mtp-box-highlight-success {
-    background: $green;
-  }
-
-  @media print {
-    background: none;
-    border: 2px solid $border-colour;
-    color: $text-colour;
-  }
-}
