--- mtp_send_money/assets-src/__old__/stylesheets/views/_references.scss
@@ -1,26 +1,0 @@
-// Application-specific
-
-.mtp-reference-actions {
-  display: none;
-
-  .js-enabled & {
-    display: block;
-
-    @media print {
-      display: none;
-    }
-  }
-}
-
-.mtp-print-reference {
-  background: file-url('icon-file-download.png') no-repeat scroll 0 0;
-  display: inline-block;
-  line-height: 40px;
-  min-height: 2.5em;
-  padding-left: 2.5em;
-
-  @include device-pixel-ratio {
-    background-image: file-url('icon-file-download-2x.png');
-    background-size: 30px 39px;
-  }
-}
