+++ mtp_send_money/assets-src/app.scss
@@ -7,5 +7,6 @@
 @import './components/filtered-list';
 @import './components/help-area';
 @import './components/prisoner-number';
+@import './components/reference-page';
 @import './components/service-charge';
 @import './components/user-agreement';
