+++ mtp_send_money/assets-src/components/reference-page/_index.scss
@@ -1,0 +1,7 @@
+.mtp-reference-actions {
+  display: none;
+
+  .js-enabled & {
+    display: block;
+  }
+}
