+++ mtp_send_money/assets-src/components/reference-page/index.js
@@ -1,0 +1,12 @@
+'use strict';
+
+export var Reference = {
+  init: function () {
+    $('.mtp-reference-actions__print').click(function (e) {
+      e.preventDefault();
+      try {
+        window.print();
+      } catch (e) {}  // eslint-disable-line
+    });
+  }
+};
