+++ mtp_send_money/templates/send_money/debit-card-on-hold.html
@@ -6,18 +6,23 @@
 
 {% block content %}
 
-  <div class="govuk-box-highlight mtp-box-highlight-info">
-    <h1 class="govuk-heading-xl">{% trans 'This payment is being processed' %}</h1>
-    <p>{% trans 'Reference:' %} <strong>{{ short_payment_ref }}</strong></p>
+  <div class="govuk-panel mtp-panel--information">
+    <h1 class="govuk-panel__title">{% trans 'This payment is being processed' %}</h1>
+    <div class="govuk-panel__body">
+      {% trans 'Your reference is' %}
+      <br />
+      <strong>{{ short_payment_ref }}</strong>
+    </div>
   </div>
 
-  <div class="mtp-reference-actions">
-    <a class="mtp-print-reference" href="#">{% trans 'Print this page' %}</a>
+  <div class="mtp-reference-actions govuk-!-margin-bottom-6 govuk-!-display-none-print">
+    <a class="mtp-reference-actions__print" href="#">{% trans 'Print this page' %}</a>
   </div>
 
   <div class="govuk-grid-row">
     <div class="govuk-grid-column-two-thirds">
-      <div class="govuk-inset-text mtp-account">
+
+      <div class="govuk-inset-text">
         <table class="govuk-table">
           <caption class="govuk-visually-hidden">
             {% trans 'Details of your payment' %}
@@ -45,6 +50,7 @@
 
       <p>{% trans 'We’ll email you a payment update in a few days.' %}</p>
       <p>{% trans 'Please don’t contact us until you have it.' %}</p>
+
     </div>
   </div>
 {% endblock %}
